backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pymupdf>=1.24.0
transformers>=4.44.0
torch>=2.2.0
sentencepiece>=0.2.0
numpy>=1.26.0
python-multipart>=0.0.9
beautifulsoup4>=4.12.3
diskcache>=5.6.3
easyocr>=1.7.0
opencv-python-headless>=4.10.0.84
Pillow>=10.2.0
httpx>=0.27.0